requirements.txt
# requirements.txt

streamlit
python-dotenv
langchain
langchain-core
langchain-openai
langchain-google-genai
langchain-community
pydantic

# For document loading
unstructured
python-docx
pypdf

# For vector storage
faiss-cpu

# For DOCX export
python-docx